+++ src/App/FAQ/FAQ.module.scss
@@ -1,6 +1,6 @@
 .faq {
   margin: 0;
-  background-color: #fafafa;
+  background-color: #f5f5f5;
 }
 
 .faqWrapper {
