--- src/App/FooterCard/FooterCard.module.scss
@@ -1,38 +1,0 @@
-.footerCard {
-  margin: 0;
-  padding: 40px 0;
-}
-
-.footerCardWrapper {
-  width: 90%;
-  max-width: 1200px;
-  margin: 0 auto;
-}
-
-.card {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  > h1 {
-    width: 70%;
-  }
-}
-
-.footerBtn {
-  cursor: pointer;
-  font-weight: 600;
-  font-size: 18px;
-}
-
-@media only screen and (max-width: 768px) {
-  .card {
-    flex-direction: column;
-    text-align: center;
-    h1 {
-      width: 90%;
-    }
-  }
-  .footerBtn {
-    margin-top: 20px;
-  }
-}
